refactor(core): move autonomy state out of tools layer
